@@ -1,6 +1,6 @@
 {
   "name": "solution-assistant",
-  "version": "0.9.3",
+  "version": "0.9.65",
   "description": "PowerApps Solution Assistant.",
   "author": "Paul Breuler",
   "license": "MIT ",
@@ -33,14 +33,14 @@
     "@material-ui/icons": "^1.1.0",
     "chartist": "0.11.0",
     "classnames": "^2.2.6",
-    "electron-is-dev": "^0.3.0",
-    "electron-log": "^2.2.16",
+    "electron-is-dev": "^1.1.0",
+    "electron-log": "^4.0.4",
     "immutability-helper": "^2.8.0",
     "nedb": "^1.8.0",
-    "perfect-scrollbar": "1.4.0",
-    "react": "^16.11.0",
+    "perfect-scrollbar": "1.5.0",
+    "react": "^16.12.0",
     "react-chartist": "0.13.1",
-    "react-dom": "^16.11.0",
+    "react-dom": "^16.12.0",
     "react-popper": "^0.10.4",
     "react-redux": "^5.1.2",
     "react-router-dom": "4.3.1",
@@ -48,14 +48,14 @@
     "react-swipeable-views": "0.12.14",
     "react-treeview": "^0.4.7",
     "recompose": "^0.30.0",
-    "redux": "^4.0.1",
+    "redux": "^4.0.5",
     "redux-thunk": "^2.3.0",
     "simple-git": "^1.107.0",
     "xml2js": "^0.4.19"
   },
   "scripts": {
-    "electron-dev": "npm run webpack:development && concurrently \"npm start\" \"wait-on http://localhost:3000 && electron .\"",
-    "electron-pack": "electron-builder -c.extraMetadata.main=build/electron.js",
+    "electron:dev": "npm run webpack:development && concurrently \"npm start\" \"wait-on http://localhost:3000 && electron .\"",
+    "electron:pack": "electron-builder -c.extraMetadata.main=build/electron.js",
     "preelectron-pack": "yarn build",
     "webpack:development": "webpack --config ./webpack.development.config.js",
     "webpack:production": "webpack --config ./webpack.production.config.js",
@@ -67,6 +67,7 @@
     "electron": "electron ."
   },
   "devDependencies": {
+    "7zip": "0.0.6",
     "@types/react": "^16.9.19",
     "@types/react-dom": "^16.9.5",
     "concurrently": "^3.6.1",
